[로또] step2: LottoMachine
