+++ src/main/java/step2/lotto/model/LottoMachine.java
@@ -15,6 +15,15 @@ public class LottoMachine {
         }
     }
 
+    public LottoTicketList buyLottoTickets(Money purchaseAmount) {
+        Count ticketCount = new Count(purchaseAmount.value() / LottoTicket.PRICE.value());
+        List<LottoTicket> lottoTickets = new ArrayList<>();
+        for (int i = 0; i < ticketCount.value(); i++) {
+            lottoTickets.add(issueLottoTicket());
+        }
+        return new LottoTicketList(lottoTickets);
+    }
+
     public LottoTicket issueLottoTicket() {
         Collections.shuffle(allLottoNumbers);
         List<LottoNumber> lottoTicket = new ArrayList<>();
